Optimisation on the initial implementation of progress_infection

+++ src/shared/types.rs
@@ -1,1 +1,1 @@
-pub type TimeStep = u16;
+pub type TimeStep = u32;
